Move remove_virtualenv to uv-fs
