+++ crates/uv-virtualenv/Cargo.toml
@@ -34,6 +34,3 @@ owo-colors = { workspace = true }
 pathdiff = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-self-replace = { workspace = true }
